requirements.txt
flask
pymongo
python-dotenv
cloudinary
torch
torchvision
Pillow